chore: CQRS.Tests renamed to Unit.Tests
