+++ Source/Tests/Unit.Tests/CQRS/UserTests.cs
@@ -16,7 +16,7 @@ using Moq;
 using Xunit;
 using static CQRS.Tests.Extensions.MockExtensions;
 
-namespace CQRS.Tests;
+namespace CQRS.Tests.CQRS;
 
 public class UserTests
 {
